requirements.txt
aws-cdk-lib>=2.38.0
constructs>=10.1.78
